// src/chain/chain.enum.ts
/**
 * ChainId
 */
export enum ChainId {
  ETH_MAINNET = '1',
  ETH_SEPOLIA = '11155111',
  ETH_GOERLI = '5',
  BASE_MAINNET = '8453',
  BASE_SEPOLIA = '84532',
  BASE_GOERLI = '84531',
  POLYGON_MAINNET = '137',
  POLYGON_MUMBAI = '80001',
  POLYGON_ZKEVM_MAINNET = '1101',
  POLYGON_ZKEVM_TESTNET = '1442',
  OPTIMISM_MAINNET = '10',
  OPTIMISM_SEPOLIA = '11155420',
  ARBITRUM_ONE = '42161',
  ARBITRUM_SEPOLIA = '421614',
  ARBITRUM_NOVA = '42170',
  BSC_MAINNET = '56',
  BSC_TESTNET = '97',
  FTM_OPERA = '250',
  FTM_TESTNET = '4002',
}